Update 2.5D preview and separate from 3D preview
